refactor: API와 관련된 코드 위치 변경
- API와 관련된 유틸리티 함수 및 타입 정의 파일 apis 디렉토리 내부로 이동
