+++ src/apis/type.ts
@@ -1,0 +1,1 @@
+export interface FetchParams { controller: AbortController }
